requirements.txt
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.3.0
pygame==2.5.2
imutils==0.5.4
